Move filters code into FilterHelper
Move the functions under filters module to FilterHelper.py

Closes https://github.com/coala/coala/issues/4494
