--- coalib/parsing/filters/__init__.py
@@ -1,4 +1,0 @@
-"""
-This package holds filter functions. Filter objects are used to
-get a list of bears with specific properties.
-"""
